refactor: front end components
