+++ web/app/components/index.tsx
@@ -1,0 +1,13 @@
+export { Button } from "./Button";
+export { DataSection } from "./DataSection";
+export { Input } from "./Input";
+export { RootSection } from "./RootSection";
+export { ConfigsModal } from "./Validations/Configurations/Configurations";
+export { ValidationAccordion } from "./Validations/Information";
+export {
+  EmptyBlock,
+  HexBlock,
+  Paragraph as P1,
+  PropBlock,
+  TextArea,
+} from "./constructors";
